add network and schedulder

+++ src/main/scala/taranis/events/Spike.scala
@@ -1,3 +1,3 @@
 package taranis.events
 
-final case class Spike(weight: Int)
+final case class Spike(weight: Double)
